Address point 1 and move the core dumping code to a new module
