+++ kernel/ukvm/platform.c
@@ -44,8 +44,8 @@ uint64_t platform_mem_size(void)
 
 void platform_dump_core(void *regs, size_t len)
 {
-    volatile struct ukvm_dump_core info;
-    memset((void *)&info, 0, sizeof(struct ukvm_dump_core));
+    volatile struct ukvm_dumpcore info;
+    memset((void *)&info, 0, sizeof(struct ukvm_dumpcore));
 
     if (len && len <= UKVM_HYPERCALL_MAX_DUMP_INFO_SIZE) {
         memcpy((void *)&info.data, regs, len);
