+++ ukvm/ukvm-configure
@@ -34,7 +34,7 @@ to build ukvm.
 
 Options:
     SOURCEDIR is the path to the ukvm source
-    MODULES can be any combination of: net blk gdb
+    MODULES can be any combination of: net blk gdb dumpcore
 
 To use a cross-compiler, set the UKVM_CC environment variable.
 To build ukvm statically, set the UKVM_STATIC environment variable.
