+++ ukvm/ukvm.h
@@ -54,8 +54,6 @@ struct ukvm_hv {
 void ukvm_elf_load(const char *file, uint8_t *mem, size_t mem_size,
         ukvm_gpa_t *p_entry, ukvm_gpa_t *p_end);
 
-/* Dump the core into the core file */
-void ukvm_dump_core(struct ukvm_hv *hv, struct ukvm_dump_core *info);
 
 /*
  * Check that (gpa) and (gpa + sz) are within guest memory. Returns a host-side
@@ -133,22 +131,6 @@ int ukvm_core_register_vmexit(ukvm_vmexit_fn_t fn);
 extern ukvm_vmexit_fn_t ukvm_core_vmexits[];
 
 /*
- * Get guest registers for core dump.
- */
-int ukvm_hv_get_regs(struct ukvm_hv *hv);
-
-/*
- * Dump the notes to the core file
- */
-int ukvm_hv_dump_notes(int core_fd, struct ukvm_hv *hv,
-        struct ukvm_dump_core *info);
-
-/*
- * Return note size
- */
-size_t ukvm_hv_get_notes_size(int *num_notes);
-
-/*
  * Module definition. (name) and (setup) are required, all other functions are
  * optional.
  */
@@ -168,6 +150,7 @@ extern struct ukvm_module ukvm_module_core;
 extern struct ukvm_module ukvm_module_blk;
 extern struct ukvm_module ukvm_module_net;
 extern struct ukvm_module ukvm_module_gdb;
+extern struct ukvm_module ukvm_module_dumpcore;
 
 /*
  * GDB specific functions to be implemented on all backends for all
@@ -228,4 +211,5 @@ int ukvm_gdb_add_breakpoint(struct ukvm_hv *hv, gdb_breakpoint_type type,
 int ukvm_gdb_remove_breakpoint(struct ukvm_hv *hv, gdb_breakpoint_type type,
                                ukvm_gpa_t addr, size_t len);
 
+ssize_t pread_in_full(int fd, void *buf, size_t count, off_t offset);
 #endif /* UKVM_H */
