+++ ukvm/ukvm_guest.h
@@ -268,7 +268,7 @@ struct ukvm_poll {
 /* UKVM_HYPERCALL_DUMP_CORE */
 #define UKVM_HYPERCALL_MAX_DUMP_INFO_SIZE 100
 
-struct ukvm_dump_core {
+struct ukvm_dumpcore {
     /* IN */
     char data[UKVM_HYPERCALL_MAX_DUMP_INFO_SIZE];
     size_t len;
